+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [0.10.3]
 
 * Refactor and simplify
+* typing of BotData
 
 # [0.10.2]